Rename file to match gem name
